extrapolate writing functions
