+++ cliRoutes/ReactComponent/classComponent.js
@@ -1,46 +1,21 @@
-const fs = require('fs')
+const writeUtils = require('./writeUtils')
 
-//React Component Custom Boilerplate
-const componentCode = require('../../cliModel/starter-code/reactComponent')
+module.exports = generateClassComponent
 
-module.exports = (
-  component,
-  componentName,
-  folderName,
-  propTypingBool,
-  reactRouterBool,
-  reduxBool
-) => {
-  if (['.', ''].includes(folderName)) {
-    const writeStream = fs.createWriteStream(`./${componentName}.js`)
-    const fileData = componentCode(
-      `${component}`,
-      `${componentName}`,
-      propTypingBool,
-      reactRouterBool,
-      reduxBool
-    )
-    writeStream.write(fileData)
-    console.log(
-      `File Creation: Class component ${component} has been created successfully!`
-    )
+/**
+ * Generates a class based react component
+ * @param {object} generateInfo - information required to generate component
+ * @param {string} generateInfo.component - type of component (e.g. class or function)
+ * @param {string} generateInfo.componentName - name of component
+ * @param {string} generateInfo.folderName - name of folder
+ * @param {boolean} generateInfo.propTypingBool - add propTypes
+ * @param {boolean} generateInfo.reactRouterBool - add react-router
+ * @param {boolean} generateInfo.reduxBool - add redux
+ */
+function generateClassComponent (generateInfo) {
+  if (['.', ''].includes(generateInfo.folderName)) {
+    writeUtils.writeInCurrentDir(generateInfo)
   } else {
-    fs.mkdir(`./${folderName}`, { recursive: false }, err => {
-      if (err) throw err
-    })
-    const writeStream = fs.createWriteStream(
-      `./${folderName}/${componentName}.js`
-    )
-    const fileData = componentCode(
-      `${component}`,
-      `${componentName}`,
-      propTypingBool,
-      reactRouterBool,
-      reduxBool
-    )
-    writeStream.write(fileData)
-    console.log(
-      `File Creation: Function component ${component} in the folder ${folderName} has been created successfully!`
-    )
+    writeUtils.writeInNewDir(generateInfo)
   }
 }
